+++ fractalistic/__main__.py
@@ -17,6 +17,10 @@ import gmpy2
 ###############
 
 @extra_command(params=[])
+@option("-t", "--threads",
+    help="Number of threads to use for rendering", 
+    type=IntRange(1), 
+    default=Settings.threads)
 @option("-ls", "--load-state", 
     help="Load state from a file", 
     type=STRING)
@@ -30,20 +34,24 @@ import gmpy2
     default="blue_brown")
 @option("-p", "--decimal-precision", 
     help="Fine-tune numeric precision by specifying the desired bit length for numeric values", 
-    type=IntRange(5))
+    type=IntRange(5),
+    default=Settings.render_settings.wanted_numeric_precision)
 @option("-i", "--max-iter", 
     help="This option defines the number of iterations required to classify a point as convergent", 
-    type=IntRange(5))
+    type=IntRange(5),
+    default=Settings.render_settings.max_iter)
 @option("-s", "--size", 
     nargs=2, 
     help="Manually set the screenshots width and height. Wont work if you use -f.", 
-    type=IntRange(32))
+    type=IntRange(32),
+    default=(Settings.screenshot_size.x, Settings.screenshot_size.y))
 @option("-f", "--fit", 
     help="Fit screenshots with the canvas.", 
     is_flag=True)
 @option("-q", "--quality", 
     help="Only when using --fit, set the quality of screenshots by multiplying the original size in the terminal by the specified scaling factor.", 
-    type=IntRange(1))
+    type=IntRange(1),
+    default=Settings.screenshot_quality)
 @option("-v", "--version", 
     help="Show version number and exit", 
     is_flag=True)
@@ -51,7 +59,7 @@ import gmpy2
     help="Enable debug mode for developers.", 
     is_flag=True)
 
-def main(fit: bool, quality: float, size: tuple[int, int], default_fractal: str, default_color: str, debug: bool, decimal_precision: int, max_iter: int, version: bool, load_state: str):
+def main(fit: bool, quality: float, size: tuple[int, int], default_fractal: str, default_color: str, debug: bool, decimal_precision: int, max_iter: int, version: bool, load_state: str, threads: int):
 
     # If -v or --version is used, show version and exit
     if version:
@@ -68,21 +76,12 @@ def main(fit: bool, quality: float, size: tuple[int, int], default_fractal: str,
 
     app.settings.fit_screenshots = fit
     app.settings.debug = debug
-
-    if quality is not None:
-        app.settings.screenshot_quality = quality
-
-    if size is not None:
-        app.settings.screenshot_size = Vec(size[0], size[1])
-
-    if max_iter is not None:
-        app.settings.max_iter = max_iter
-
-    if decimal_precision is not None:
-        app.settings.render_settings.wanted_numeric_precision = decimal_precision
-
-    if load_state is not None:
-        app.settings.state_file = load_state
+    app.settings.screenshot_quality = quality
+    app.settings.screenshot_size = Vec(size[0], size[1])
+    app.settings.max_iter = max_iter
+    app.settings.render_settings.wanted_numeric_precision = decimal_precision
+    app.settings.state_file = load_state
+    app.settings.threads = threads
 
     app.run()